Moved images under e2e package, and removed Cortex images.
