+++ integration/backward_compatibility_test.go
@@ -14,17 +14,18 @@ import (
 	"github.com/cortexproject/cortex/integration/e2e"
 	e2edb "github.com/cortexproject/cortex/integration/e2e/db"
 	"github.com/cortexproject/cortex/integration/e2ecortex"
-	"github.com/cortexproject/cortex/integration/images"
 )
 
 var (
+	// If you change the image tag, remember to update it in the preloading done
+	// by CircleCI too (see .circleci/config.yml).
 	previousVersionImages = []string{
 		// 0.6.0 used 204 status code for querier and ingester
 		// distributor didn't have /ready page, and we used check on the /ring page instead
-		images.Cortex0_6_0,
+		"quay.io/cortexproject/cortex:v0.6.0",
 
 		// 0.7.0 used 204 status code for all components
-		images.Cortex0_7_0,
+		"quay.io/cortexproject/cortex:v0.7.0",
 	}
 )
