+++ integration/e2e/db/db.go
@@ -9,7 +9,7 @@ import (
 	awscommon "github.com/weaveworks/common/aws"
 
 	"github.com/cortexproject/cortex/integration/e2e"
-	"github.com/cortexproject/cortex/integration/images"
+	"github.com/cortexproject/cortex/integration/e2e/images"
 )
 
 const (
