+++ integration/e2e/images/images.go
@@ -1,11 +1,10 @@
 package images
 
-// If you change the image tag, remember to update it in the preloading done
-// by CircleCI too (see .circleci/config.yml).
-const (
-	Cortex0_6_0 = "quay.io/cortexproject/cortex:v0.6.0"
-	Cortex0_7_0 = "quay.io/cortexproject/cortex:v0.7.0"
+// These are variables so that they can be modified.
 
+var (
+	// If you change the image tag, remember to update it in the preloading done
+	// by CircleCI too (see .circleci/config.yml).
 	Memcached        = "memcached:1.6.1"
 	Minio            = "minio/minio:RELEASE.2019-12-30T05-45-39Z"
 	Consul           = "consul:0.9"
